pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "open-retrosight"
version = "0.1.0"
description = "非侵入式工业边缘AI网关 - 给老机器装上数字眼睛"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Open-RetroSight Team"}
]
keywords = ["industrial", "iot", "ocr", "computer-vision", "edge-computing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Manufacturing",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    "opencv-python-headless>=4.5.0,<5.0.0",
    "numpy>=1.20.0,<2.0.0",
    "paddlepaddle>=2.4.0,<3.0.0",
    "paddleocr>=2.6.0,<3.0.0",
    "paho-mqtt>=1.6.0,<2.0.0",
    "streamlit>=1.20.0,<2.0.0",
]

[project.optional-dependencies]
pointer = [
    "ultralytics>=8.0.0,<9.0.0",
]
modbus = [
    "pymodbus>=3.0.0,<4.0.0",
]
dev = [
    "pytest>=7.0.0,<8.0.0",
    "pytest-cov>=4.0.0,<5.0.0",
    "pytest-asyncio>=0.21.0,<1.0.0",
    "black>=23.0.0,<25.0.0",
    "ruff>=0.1.0,<1.0.0",
]
all = [
    "open-retrosight[pointer,modbus,dev]",
]

[project.urls]
Homepage = "https://github.com/open-retrosight/open-retrosight"
Documentation = "https://github.com/open-retrosight/open-retrosight/tree/main/docs"
Repository = "https://github.com/open-retrosight/open-retrosight"

[project.scripts]
retrosight = "retrosight.ui.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
